Updated wire channel cutout
My original idea had the wires running under the belt but that bundle of wires is big enough to lift the pads away from our skin. New design is going to run the wires directly below the belt. May need to design and print a few wire harness holders for the belt.