Bigly bugs
Got the first hardware prototype of the belt finished (barely) but when I tested it out I got a ton of random vibrating all over the place. Nothing worked right lol. They say in the military that no plan survives first contact with the enemy. Well it seems no code (of mine) survives first contact with the hardware! In my defense these are classic C-lang bugs and I just dont work with C often enough.

The very first (and probably the most severe) bug was that I was using the motorpin value as an index, which was very bad. I decided to just track the index in a variable but I'm thinking just as I type this that I could possibly just use a pointer instead, and then I could increment the pointer with no index tracking variable. Maybe in the future if this fix doesn't work.

The next issue is the PWM range. Ideally (and perhaps naively), if we wanted to create a virtual vibrating motor between two adjacent motors, we would just set both motors to 128. Or maybe set both to 255, always having one motor at max while the other is faded out. At any rate the issue is that the motors don't respond in a linear fashion with respect to the PWM value. There is no activity until the PWM gets around ~120. That means the virtual nodes will be almost not even active at all.

So yeah a few little issues and the hardware is extremely fragile. My first prototype design I just straight up wired every motor but with a single ground wire for all. That means 7 wires at the thickest point in the harness where it leaves the belt. The wires on the motors are very easy to break off so I add to hotglue them into the wire channel, which seems to work fine. The second fragile point is the point where the ground wire of the motor connects to the common ground, the solder joint there gets worked a lot and one has already broken loose.

This program when working, should make it feel like there is a single vibration moving around the circumference of the belt based on the angle of the potentiometer. Once I can get that somewhat correct I can move onto filtering the heading information.