Actual bug fix for real
Last update was just a code review. Actually running the code on the hardware was still a hot mess. To get things sorted out I swapped the belt for 6 LEDs as an output and it made diagnosing the issue so much easier. First off, the fix for the motor array index was itself flawed with the very issue it was supposed to fix! Instead of setting the index at line 65 (67 now), I was setting the pin number. Doh! The check of the index to roll over if we got larger than the array size was also flawed, somehow forgetting the array is indexed from 0. I also had the ratios backwards for motor1 and motor2, so just doing a 1-ratio solved that easily. Finally the debug println showing PWM vals was only showing motor1 for both values. Copypasta lyphe.