Better PWM
Forgot to mention in the last commit that I also changed the way the PWM values were being written because the motors were staying on after they were set to any PWM value. At first I decided to just delay after writing the PWM and then write a 0 to the 2 active motors. That worked but didn't produce a continuous output. New solution is to write a 0 first to all motors and then the PWM value and that provides a perfectly continuous and smooth output. At this point I think I got this test code working and now I can get the heading information sorted out.

The only two issues left I can think of atm is that the 360 degree point is actually overshooting and landing on the first motor and not the last motor. The 0 degree point is correctly on the first motor. The second issue is a weird flicker in the LED test output at low PWM values. But whenever I enable debug prints the issue goes away. Best theory now is that the serial printing causes a delay which keeps the PWM signal on longer. So I added in a small delay and it greatly helps but the flickers is still sorta there.