Filtering heading
Realized that I didn't need to add up every heading in the buffer every time I wanted a new average, I just had to keep a sum and subtract from it the oldest value in the buffer while adding the newest value. Kind of a poor-mans PID but it works reasonably well for what I need. Changing the value of buffer_size will affect how slowly the value changes. Current value of 25 seems to take the filter about a second to catch up to any massive heading changes. Will tune this further when real heading data is being input.