Changed test mode
Instead of the pot controlling the angle of the output, it now controls the speed at which the output is cycled from 0-360.  This mode allowed me to dial in certain parameters more easily, like the pwm value. This produces the smoothest output so far although you can feel small gaps every now and then.

Still have the major crashing issue with the arduino. Even just outputting the heading without any parts of the motor program is crashing eventually, which the sensormonitor app never did (or I never experienced it). Not sure if its the hardware or what because the sensormonitor app ran on an arduino nano and this is running on a very old uno. I may switch over to an mega2560 I have and see if the issue continues. Maybe I just have a bad board.