Works, using mega 2560 now
I was hoping I wouldnt have to resort to the mega because of its size but so far everything is now working great. The excess of PWM pins is a nice thing to have as well since before we had used all 6. That allowed me to add back in the status indicator LEDs.

One big change was to the order of the motor pins. In the old order everything was backwards with north on your back. I just moved everything over 3 slots and now its aligned properly. Besides that its the same code, just now it doesnt crash. I still need to solve the power issue by just not driving the motors directly from the pins. Will look into if I can get away with some simple npn transistors, been a while since elec 201.

As far as the overall device performance, its a bit lacking. One part of that is that I dont have filtering working and that messy signal is very annoying. However when the compass data coming in is good and steady, the device works surprisingly well. I'm almost ready to miniaturize everything and put it in a nice case. Unfortunately the mega is going to make that job a lot more difficult. I almost want to just get a simple barebones atmega 2560 schematic layout and make my own board in eagle or something. Protoboard and lots of jumper wire seems to be in my immediate future however lol.