Testing a very basic fix for filtering
I wanted to go the full PID route but decided to try and just transform the angle we were getting with a period function. I settled on tan/atan2 since it seemed like the only way to get a continuous output in the range I needed. I had to track both the sine and cosine of our heading angle though so thats not great. So although the maths are a little brute-forced, it does actually work. It only goes haywire when there are massive shifts in the angle, like +- 90 degrees but the values eventually settle again.