Fixed bug in heading filter
Fixed a major bug where the cosine parts of the filter were actually being fed sine data. It kind of sort of actually worked even with the bug but the weird issues I was having are seemingly just this bit of code. Moved the setup() function to an easier to find spot. I should really standardize on a single way to order functions.

Next up for the project is miniaturizing this whole thing so I can start walking around with it full time. I bought an $11 "AITRIP Mega 2560 Pro Embed". Its a tiny (3.8cm x 5.5cm) atmega2560 board with full pinout. This thing is going to be sweet and small, should fit really nice in a sleek case I can have on the belt or in my pocket. The SEN0140 absolutely has to be on the belt though so thats an interesting one to think about.