First hardware prototype
Finished soldering up the first hardware prototype and it seems like it actually worked on the first try! That like almost never happens lol, I usually end up resolding one or two things. I made doubly sure my breadboard circuit worked and then just copied it all over. I didn't include the potentiometer so Its been disabled in the code for now. Also I didn't solder the sensor onto anything yet as I'm still undecided where to locate it.

The first instinct is just to put the sensor on the same board as everything else and jam it in our pocket. Thing is, our tilt compensation aint the greatest so the random orientation in our pocket probably won't be great. The second idea was to have it on the belt with a cable attaching to the brainbox. The third idea is kind of a mix of the two. We put the entire brainbox on the belt and have the sensor plug in to the box through an opening (pin header connection). I really don't want to solder a $30 sensor permanently into a crappy prototype if I can avoid it.

Next up now that I have a rough layout of the board I can start to mock up a case for it and get to printing. Then figure out the sensor and finally add in the status LED and potentiometer (maybe) and its done! Really the most iffy part of this entire operation was the power hardware. The crashing I was running into on the mega and uno boards appears to be power related. A document I was reading seemed to indicate that powering the arduino via USB limits the current draw to 400mA instead of the 800mA normally stated. Now the motors are powered independently directly from the USB connector and we're just using 2N2222 transistors to switch them on and off. I wasn't sure if they would play well with PWM signals but all appears to be fine.