Added corrections for sensor heading misalignment
First trials with the prototype hardware I found immediately that I had to have the sensor perfectly aligned with my body for the whole thing to work properly. This correction offset will fix that. At the end of magcal or maybe on button press (or both?) we will give the wearer a few seconds to align their body to north using another compass. We then take the difference between that heading and our sensor's heading and get an offset we can add or subtract. This should allow the sensor to be placed in an arbitrary orientation and still function properly. Im still a bit worried that at extreme tilt angles my questionable tilt compensation maths are going to stop being reliable.