Sub-board brainbox half started
Got the side of the case holding the sub board basically designed. Got most of the way complete when I realized there was no space for case screw holes so I had to improvise with the bulbs on the corners. I also think I may need to add a removable brace behind the USB port since its not mechanically fixed to the PCB.