Software adaptations to hardware limitations
Decided just to leave the current hardware alone and use it for a while. To that end instead of a status LED I decided to use vibration patterns instead. 2 slow buzzes means calibration has started, 2 fast buzzes indicates we are about to take the compass offset heading (so we need to face north). Currently it waits 5 seconds before it takes the offset reading. Finally 3 slow buzzes means everything is ready to go.