Starting work on custom PCB
I'm tired of making PCBs by hand using crappy protoboard and the constraints it brings. I have a laser printer and a drill already so with just some ferric chloride and raw PCB blanks I should be able to crank out something that looks nice and is easy to work with. Unfortunately that nice idea ends right where my inexperience with KiCAD begins. I have a schematic of the final circuit complete (I think) and next is to lay out the physical PCB now.

Also added 2 new potentiometers to the circuit so I had to add new references in the arduino code. The variable names are now perfectly descriptive.