Physical PCB start
Sort of got a start on the physical PCB layout. Its hard to lay things out in ways that don't overlap with other traces on a single layer. I got most of everything connected up and I think what I have left will need to be on the other side of the PCB. Will try and figure that out tomorrow.