PCB version 1.0
Ok I've finished laying everything out. In the end I was successful in laying out the data and power traces on top and only have ground traces on the bottom. I've generated the reversed masks that are ready for laser printing. Now I'm realizing that transferring the mask to the PCB is going to be a slightly involved task. I've settled on the rolling pin + hotplate method since I don't want to build a jig to hold an iron upside down and I don't want to spend the time modifying a cheap ebay laminator to work with PCBs. Going forward I may switch to a photoetching process instead since that doesn't require heat+pressure. Although it requires a few more chemicals it seems like a more controllable process. I did price out an online mfg and it was around $50 shipped for 5 boards. I'd be amazed if the first version of this board actually even works but maybe a future version I will order online instead.

Also something happened with the symbol library for the project and it generated a rescue lib. No idea why but it seems fine.