Size matters
Welp Got all the way to putting the transfer paper onto the PCB when I found out that my raw PCBs are kinda small. 70mm x 100mm - which is a lot smaller than the current footprint of like 90mm x 120mm. So LOTS of rearranging things and reassigning pins. I think the end result is a far better board than the previous one so It was well worth the effort. I made sure to not route any traces too close to the edge, other traces, or other pin holes. Trying to give the toner transfer the best possible chance of at least not bridging stuff.