Rerouting to be single-sided
Since I can't plate my through-holes, any connections on the opposite side of the board had to be soldered too. The problem there is that the components are in the way so you can't solder. Because of this I decided to route all the ground traces on the top. At first I had planned to just put some through-hole vias that were away from components to fix that soldering issue. Then I realized I could route them all on the top no problem, so I went that way. The traces are really the "bottom" now and the top is where the components are with the project name/version mask.

Also didn't realize I forgot to connect one of the pins for the SEN0140 board. I guess I deleted the trace with the intent to reroute it but never did. Fixed now.

Finally, all my hard work with the last board was for nothing in the end as I had my print settings at the default. Apparently microsoft edge defaults to printing everything scaled instead of 100%. Got all finished drilling the holes by hand with a pin vice and then found the headers wouldnt line up. ggwp git gud. :(