Updated readme and changed to public
Taking a break from this project for a little while so I figured I would make it public. Most of the work is done. The code is a little shaky in terms of stability (and I'm not a C coder so its not surprising) and the hardware needs some changes to work properly. The base idea is sound though and I have tested the device several times successfully. I havent been able to wear it all the time because of the stability issues.

I'm considering maybe switching from cellphone vibration motors to direct current stimulation of the skin. I have a TENS device for an old injury that got me thinking how it could be repurposed for tactile stimulation. Its extremely effective and is also very widely tunable in terms of the intensity and feel of the pulses (by modulating amplitude and frequency, maybe even the waveform could have effects). The only issue there is that it probably requires using those gel pads. Not sure you could get the same sensitivity and consistency by using just skin contact without some adhesive.